Add a config option $WARN_HEADERS to only warn when the $ HEADER $ isn't present or corrupted, instead of bombing out.
Submitted by:	Mathieu Arnold <arn_mat@club-internet.fr>

+++ CVSROOT/commit_prep.pl
@@ -286,9 +286,8 @@ if ($cfg::CHECK_HEADERS) {
 
 		# Unexpand the rcsid if required.
 		fix_up_file($arg) if $cfg::UNEXPAND_RCSID and !$failed;
-
 	}
-	if ($failed) {
+	if ($failed and not $cfg::WARN_HEADERS) {
 		print "\n";
 		unlink($cfg::LAST_FILE);
 		exit(1);
